requirements.txt
absl-py==1.4.0
accelerate==0.28.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
bitsandbytes==0.43.0
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cmake==3.28.3
comm==0.2.1
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.34
datasets==2.18.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
dpcpp-cpp-rt==2024.0.2
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fasttext==0.9.2
filelock==3.9.0
fonttools==4.41.1
frozenlist==1.3.3
fsspec==2023.5.0
gmpy2==2.1.2
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
huggingface-hub==0.22.2
idna==3.4
imgcat==0.5.0
importlib-metadata==6.7.0
importlib-resources==6.0.0
intel-cmplr-lib-rt==2024.0.2
intel-cmplr-lic-rt==2024.0.2
intel-opencl-rt==2024.0.2
intel-openmp==2024.0.2
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.0
jupyter-client==7.3.4
jupyter_core==5.7.1
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0
lit==17.0.6
llvmlite==0.40.0
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl==2024.0.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
munkres==1.1.4
nest_asyncio==1.6.0
networkx==2.8.4
nltk==3.8.1
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.2
pandas==2.0.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
piqa==1.3.2
platformdirs==4.2.0
pooch==1.6.0
prompt-toolkit==3.0.42
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==23.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0
pytorch-msssim==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
setuptools==66.0.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
sympy==1.11.1
tbb==2021.11.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.1
tqdm==4.65.0
traitlets==5.14.1
transformers==4.39.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2023.3
unicodedata2==15.0.0
urllib3==1.26.15
wcwidth==0.2.13
Werkzeug==2.3.6
wget==3.2
wheel==0.38.4
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
